+++ CHANGELOG.md
@@ -8,4 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`88cf670...main`][88cf670...main].
 
+### Added
+
+- Added `Markdown\VersionCollector` ([#12]), by [@localheinz]
+
 [88cf670...main]: https://github.com/ergebnis/keep-a-changelog/compare/88cf670...main
+
+[#12]: https://github.com/ergebnis/keep-a-changelog/pull/12
+
+[@localheinz]: https://github.com/localheinz